home/hero.html: fix download link
Prior to v2.5.0, Git LFS named packaged release artifacts as:

    git-lfs-darwin-amd64-2.4.0.tar.gz

and etc., for combinations of OS, architecture, version, and archival
format.

In v2.5.0, we began naming those same release artifacts as:

    git-lfs-darwin-amd64-v2.5.0.tar.gz

Noting that there is now a 'v' prefix for the version number. In lieu of
changing this back to the sans-'v' format, let's keep this and continue
to name versions as-is from here going forward.

This will help alleviate some whiplash experienced by packaging
maintainers who have updated scripts to account for this new name, and
it will be easier to maintain in Git LFS's somewhat-complicated
Makefile.

So, given that this change is here to stay (and we only ever increase
the version number ;-) ), let's update the link on the website to
download the appropriate thing, and not 404.