Clarify when/where to run `git lfs install`
Previous instructions are vague on where the command should be run (eg
it needs to be *in* the repo directory so it can install hooks) and
misleading in that it suggests you only need to set up Git LFS once.
What it means to say is that you only need to set it up once, _per
repository_ as the relevant hooks need to be installed per-repo.

This clarifies both points and fixes a minor display issue where the
Download link actually incorporates the whitespace after the word
"Download".