_includes: use zip files for macOS in second link
In commit 12d0190fbdd75e687cec73fe92c4e2aead768db0 the main
"hero" download link was updated to offer a Zip archive file
link for macOS instead of a gzipped tarball, and a secondary
link was also updated in that same commit.  That link was
subsequently removed.

However, another secondary link which still exists was never
updated, so we change it now to offer the Zip archive file.