_includes,js: revise and add download buttons
With the advent of Apple's M1-based systems we now need to offer
two download options for macOS users on our home page, as there is
no convenient and reliable way to distinguish between hardware
platforms in JavaScript.

Therefore we add a second "Apple Silicon" download button for
macOS users, and rename the original button "Intel Silicon".

Rather than try to continue the existing practice of sharing the
HTML code for the first download button with Windows, we refactor
the download button HTML into three sections, one each for Windows,
macOS, and Linux.  This in turn allows us to delete the "js-os-data"
CoffeeScript as each HTML section is now fully independent.

Having made this change for macOS, it also makes some sense for us
to make a parallel change for Linux users, promoting the Intel/AMD
x86-64 download option to a button aligned below the packagecloud.io
install option button.

Because we now have two primary download options for macOS, and for
Linux as well, we drop the hyperlink around the "Download" text in
the our set of instructions for new users, as it is not useful to have
that text always link to just one of several possible download options.

Lastly we add some simple whitespace-based padding just to make some
of our text strings align slightly better in both the macOS and Linux
versions of the home page.